+++ tests/bootstrap.php
@@ -78,7 +78,8 @@ $connection->execute('
         integer_field INT,
         length_between_field VARCHAR(255),
         natural_number_field INTEGER,
-        unique_field VARCHAR(255) UNIQUE
+        unique_field VARCHAR(255) UNIQUE,
+        email_field VARCHAR(255)
     )
 ');